requirements/local.txt
-r base.txt

watchfiles==0.22.0
black==24.8.0
django-debug-toolbar==6.0.0
